+++ modules/devices.py
@@ -160,7 +160,7 @@ def test_bf16():
 
 
 def set_cuda_params():
-    shared.log.debug('Verifying Torch settings')
+    # shared.log.debug('Verifying Torch settings')
     if cuda_ok:
         try:
             torch.backends.cuda.matmul.allow_tf32 = True
@@ -212,8 +212,7 @@ def set_cuda_params():
     else:
         inference_context = torch.no_grad
     shared.log.debug(f'Desired Torch parameters: dtype={shared.opts.cuda_dtype} no-half={shared.opts.no_half} no-half-vae={shared.opts.no_half_vae} upscast={shared.opts.upcast_sampling}')
-    shared.log.info(f'Setting Torch parameters: dtype={dtype} vae={dtype_vae} unet={dtype_unet} context={inference_context.__name__} fp16={fp16_ok} bf16={bf16_ok}')
-    shared.log.debug(f'Torch default device: {torch.device(get_optimal_device_name())}')
+    shared.log.info(f'Setting Torch parameters: device={torch.device(get_optimal_device_name())} dtype={dtype} vae={dtype_vae} unet={dtype_unet} context={inference_context.__name__} fp16={fp16_ok} bf16={bf16_ok}')
 
 
 args = cmd_args.parser.parse_args()